+++ packages/stylelint-config-terra/stylelint.config.js
@@ -18,7 +18,7 @@ module.exports = {
         message: 'Custom property names should be written in lowercase with hyphens',
       },
     ],
-    'declaration-property-unit-whitelist': [{ 'line-height': [] }, { severity: 'warning' }],
+    'declaration-property-unit-allowed-list': [{ 'line-height': [] }, { severity: 'warning' }],
     'terra/custom-property-name': true,
     'terra/custom-property-no-duplicate-declaration': true,
     'terra/custom-property-pattern': true,