+++ packages/terra-functional-testing/src/commands/expect/toMatchReference.js
@@ -6,7 +6,8 @@
  * @param {boolean} screenshot.isNewScreenshot - If the latest screenshot was a new screenshot that has no reference screenshot.
  * @param {boolean} screenshot.isSameDimensions - If the latest screenshot was the same size as the reference screenshot.
  * @param {boolean} screenshot.isWithinMismatchTolerance - If the latest screenshot was within the mismatch tolerance.
- * @param {Number} screenshot.misMatchPercentage - The mismatch percentage when comparing the latest screenshot to the reference screenshot.
+ * @param {number} screenshot.misMatchPercentage - The mismatch percentage when comparing the latest screenshot to the reference screenshot.
+ * @param {boolean} screenshot.screenshotWasUpdated - If the reference screenshot was updated with the latest captured screenshot.
  * @returns {Object} - An object that indicates if the assertion passed or failed with a message.
  */
 function toMatchReference(screenshot) {
@@ -15,11 +16,12 @@ function toMatchReference(screenshot) {
     isSameDimensions,
     isWithinMismatchTolerance,
     misMatchPercentage,
+    screenshotWasUpdated,
   } = screenshot;
 
   // Validate the screenshot is the same size for the case when the new screenshot matches 100% but is 'n' pixel taller due to new content.
   // For example: the latest screenshot of a list has two more items than the reference screenshot so it is 60 pixels taller. That should fail.
-  const imagesMatch = isNewScreenshot || (isSameDimensions && isWithinMismatchTolerance);
+  const imagesMatch = isNewScreenshot || screenshotWasUpdated || (isSameDimensions && isWithinMismatchTolerance);
   let message = '';
 
   if (!isSameDimensions) {