+++ packages/terra-functional-testing/src/config/utils/getCapabilities.js
@@ -2,14 +2,14 @@ const { chrome, firefox, ie } = require('../capabilities');
 
 /**
  * Determines the browser capabilities for the WebDriver session.
- * @param {array} browsers - An array of browser names to enable. (chrome, firefox, ie)
+ * @param {string|array} browsers - A string or array containing a list of browser names. (chrome, firefox, ie)
  * @param {boolean} isGridEnabled - Whether the WebDriver session is running against the remote selenium grid.
  * @returns {array} - An array of browser capabilities.
  */
 const getCapabilities = (browsers, isGridEnabled) => {
   const capabilities = [];
 
-  if (!browsers) {
+  if (!browsers || browsers.length === 0) {
     // Enable chrome as the default browser.
     capabilities.push(chrome);
 
@@ -18,15 +18,17 @@ const getCapabilities = (browsers, isGridEnabled) => {
       capabilities.push(firefox, ie);
     }
   } else {
-    browsers.forEach((browser) => {
-      if (browser === 'chrome') {
-        capabilities.push(chrome);
-      } else if (browser === 'firefox') {
-        capabilities.push(firefox);
-      } else if (browser === 'ie' && isGridEnabled) {
-        capabilities.push(ie);
-      }
-    });
+    if (browsers.indexOf('chrome') > -1) {
+      capabilities.push(chrome);
+    }
+
+    if (browsers.indexOf('firefox') > -1) {
+      capabilities.push(firefox);
+    }
+
+    if (browsers.indexOf('ie') > -1 && isGridEnabled) {
+      capabilities.push(ie);
+    }
   }
 
   // Randomize the browser order.