+++ packages/terra-functional-testing/src/services/wdio-visual-regression-service/methods/BaseCompare.js
@@ -9,17 +9,20 @@ class BaseCompare {
    * @param {Object} options - Service configuration options.
    * @param {Object} options.locale - The locale being tested.
    * @param {Object} options.theme - The theme being tested.
+   * @param {boolean} options.updateScreenshots - Whether or not all reference screenshots should be updated even when there are mistmach with latest.
    */
-  constructor(options) {
+  constructor(options = {}) {
     const {
       locale,
       theme,
+      updateScreenshots,
     } = options;
 
     // screenshot naming config
     this.baseScreenshotDir = process.cwd();
     this.locale = locale || 'en';
     this.theme = theme || 'terra-default-theme';
+    this.updateScreenshots = updateScreenshots === true;
   }
 
   /**
@@ -40,8 +43,8 @@ class BaseCompare {
   /**
    * Creates the sanitized test name for the screenshot.
    *
-   * @param {String} fullName - The test name.
-   * @returns {String} - The test name.
+   * @param {string} fullName - The test name.
+   * @returns {string} - The test name.
    */
   createTestName(fullName) {
     const matches = TEST_ID_REGEX.exec(fullName);
@@ -117,13 +120,14 @@ class BaseCompare {
   /**
    * Creates the screenshot comparison report object.
    *
-   * @param {Boolean} referenceExists - Whether or not the screenshot was just created.
-   * @param {Number} misMatchPercentage - The percent mismatched of the latest screenshot compared to the reference screenshot.
-   * @param {Boolean} isWithinMismatchTolerance - Whether or not the latest screenshot is a close enough match the reference screenshot.
-   * @param {Boolean} isSameDimensions - Whether or not the latest screenshot was the same dimensions as the reference screenshot.
+   * @param {boolean} referenceExists - Whether or not the screenshot was just created.
+   * @param {number} misMatchPercentage - The percent mismatched of the latest screenshot compared to the reference screenshot.
+   * @param {boolean} isWithinMismatchTolerance - Whether or not the latest screenshot is a close enough match the reference screenshot.
+   * @param {boolean} isSameDimensions - Whether or not the latest screenshot was the same dimensions as the reference screenshot.
+   * @param {boolean} screenshotWasUpdated - Whether or not the reference screenshot was updated with the latest captured screenshot.
    * @returns {Object} - The relevant comparison results to report.
    */
-  createResultReport(referenceExists, misMatchPercentage, isWithinMismatchTolerance, isSameDimensions) {
+  createResultReport(referenceExists, misMatchPercentage, isWithinMismatchTolerance, isSameDimensions, screenshotWasUpdated) {
     if (!referenceExists) {
       return { isNewScreenshot: true };
     }
@@ -132,6 +136,7 @@ class BaseCompare {
       misMatchPercentage,
       isWithinMismatchTolerance,
       isSameDimensions,
+      screenshotWasUpdated,
     };
   }
 }