+++ packages/terra-functional-testing/src/services/wdio-visual-regression-service/utils/image/gm.js
@@ -40,7 +40,7 @@ async function cropImage(base64Screenshot, cropDimensions) {
 /**
  * Scales an image down or up.
  * @param {string} base64Screenshot - image to scale
- * @param {Number} scaleFactor - scale factor, e.g. 0.5 for downscale or 1.5 for upscale
+ * @param {number} scaleFactor - scale factor, e.g. 0.5 for downscale or 1.5 for upscale
  * @return {Promise} resolves to cropped image
  */
 async function scaleImage(base64Screenshot, scaleFactor) {